PAR/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.23.1
anyio==4.3.0
arxiv==2.1.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.0
feedparser==6.0.10
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.14
langchain-anthropic==0.1.6
langchain-community==0.0.31
langchain-core==0.1.38
langchain-openai==0.1.1
langchain-pinecone==0.0.3
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langgraph==0.0.32
langsmith==0.1.38
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.16.0
orjson==3.10.0
packaging==23.2
pinecone-client==3.2.2
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
tqdm==4.66.2
types-requests==2.31.0.20240402
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
wikipedia==1.4.0
yarl==1.9.4
youtube-search==2.1.2
